primer avance hecho

+++ SMDB/SMBD.cs
@@ -7,6 +7,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using SMDB.Classes;
+using System.IO;
 
 namespace SMDB
 {
@@ -22,20 +24,26 @@ namespace SMDB
 
             //Size Properties
             this.WindowState = FormWindowState.Maximized;
+            TreeView.Width = this.Width;
+            TreeView.Height = this.Height;
+            SplitContainer.Width = this.Width;
+            SplitContainer.Height = this.Height;
+
             this.MinimumSize = this.Size;
 
             //Menu Properties
             MenuSeparator1.Enabled = false;
-            MenuSeparator2.Enabled = false;
-            MenuSeparator3.Enabled = false;
             MenuSeparator4.Enabled = false;
-            MenuSeparator5.Enabled = false;
+            MenuSeparator3.Enabled = false;
+            
 
             MenuDisconnectDB.Enabled = false;
-            MenuClose.Enabled = false;
             MenuDeleteDB.Enabled = false;
+            MenuHDisconnectDB.Enabled = false;
+            MenuHDeleteDB.Enabled = false;
         }
 
+        //Decapricated
         private void Form1_Paint(object sender, PaintEventArgs e)
         {
             //Graphics G = e.Graphics;
@@ -60,35 +68,10 @@ namespace SMDB
                 MessageBox.Show("DISCONNECT FROM DATABASE");
                 return true;
             }
-            if (keyData == (Keys.Control | Keys.X)) //CLOSE
-            {
-                MessageBox.Show("CLOSE CURRENT FILE OR PROJECT");
-                return true;
-            }
-            if (keyData == (Keys.Control | Keys.N)) //NEW FILE
-            {
-                MessageBox.Show("NEW FILE");
-                return true;
-            }
-            if (keyData == (Keys.Control | Keys.O)) //OPEN FILE
-            {
-                MessageBox.Show("OPEN FILE");
-                return true;
-            }
             if (keyData == (Keys.Alt | Keys.Control | Keys.N)) //NEW DATABASE
             {
                 //MessageBox.Show("NEW DATABASE");
                 CreateDatabase();
-                return true;
-            }
-            if (keyData == (Keys.Control | Keys.Shift | Keys.N)) //NEW PROJECT
-            {
-                MessageBox.Show("NEW PROJECT");
-                return true;
-            }
-            if (keyData == (Keys.Control | Keys.Shift | Keys.O)) //OPEN PROJECT
-            {
-                MessageBox.Show("OPEN PROJECT");
                 return true;
             }
             if (keyData == (Keys.Control | Keys.Alt | Keys.X)) //DELETE DB
@@ -102,16 +85,54 @@ namespace SMDB
         public void OpenDatabase()
         {
             FolderBrowserDialog folderBrowserDialog = new FolderBrowserDialog();
-            
+            string currentAppDirectory = AppDomain.CurrentDomain.BaseDirectory;
+            string lastPath = currentAppDirectory + "\\" + "lastPath.lp";
+            FileHandler.createFile(lastPath);
+            //FileStream file_lastPath = FileHandler.openFile(lastPath);
+
             if(folderBrowserDialog.ShowDialog() == DialogResult.OK){
                 this.currentPath = folderBrowserDialog.SelectedPath;
                 string[] folders = this.currentPath.Split('\\'); 
                 this.currentDB = folders[folders.Length-1];
-
-                MenuUpdate(true);
-                TitleUpdate();
-                DisplayDB();
+                if (isDB())
+                {
+                    MenuUpdate(true);
+                    TitleUpdate();
+                    DisplayDB();
+                }
+                else
+                {
+                    string message = "The selected object is not a DB";
+                    DialogResult result = MessageBox.Show(message, "Not a DB", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    OpenDatabase();
+                }
+                
             }
+        }
+
+        public void CloseDatabase()
+        {
+            TreeView.Nodes.Clear();
+            MenuUpdate(false);
+            this.Text = "SMBD";
+            this.currentDB = "";
+            this.currentPath = "";
+        }
+
+        public bool isDB()
+        {
+            bool response = false;
+            string[] files = System.IO.Directory.GetFiles(this.currentPath);
+            foreach (string file in files)
+            {
+                string filename = file.Split('.')[1];
+                if (filename == "db")
+                {
+                    response = true;
+                    break;
+                }
+            }
+            return response;
         }
 
         public void CreateDatabase()
@@ -119,24 +140,53 @@ namespace SMDB
             this.Create_DB = new CREATE_DB();
             this.Create_DB.Parent_SMBD = this;
             this.Create_DB.Show();
-
-            MenuUpdate(true);
-            TitleUpdate();
         }
 
         public void DeleteDatabase()
         {
-            string message = String.Format("Are you sure you want to delete the current Database{0}{1}",Environment.NewLine,this.currentDB);
-            DialogResult result = MessageBox.Show(message, "ALERT", MessageBoxButtons.YesNo,MessageBoxIcon.Warning);
-            
-            if (result == DialogResult.Yes)
+            string messageAnswer = String.Format("Are you sure you want to delete the current Database{0}{1}",Environment.NewLine,this.currentDB);
+            string messageError = "The object is not available";
+            if (isDB())
             {
-                System.IO.Directory.Delete(this.currentPath, true);
-                this.currentPath = "";
-                this.currentDB = "";
-                this.Text = "SMBD";
-                MenuUpdate(false);
+                DialogResult result = MessageBox.Show(messageAnswer, "Delete Database!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+
+                if (result == DialogResult.Yes)
+                {
+                    System.IO.Directory.Delete(this.currentPath, true);
+                    this.currentPath = "";
+                    this.currentDB = "";
+                    this.Text = "SMBD";
+                    MenuUpdate(false);
+                    TreeView.Nodes.Clear();
+                }
             }
+            else
+            {
+                MessageBox.Show(messageError, "Invalid Object", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+        }
+
+        public void RenameDatabase(string newName)
+        {
+            string origin = this.currentPath;
+            string destiny = this.currentPath.Replace(this.currentDB,newName);
+            string[] files = Directory.GetFiles(origin);
+            string[] newFiles = new string[files.Length];
+            for (int i = 0; i < files.Length; i++)
+            {
+                string newFile = files[i].Replace(this.currentDB,newName);
+                newFiles[i] = newFile;
+            }
+            if (isDB())
+            {
+                Directory.Delete(origin, true);
+                Directory.CreateDirectory(destiny);
+                FileHandler.createFiles(newFiles);
+                this.currentPath = destiny;
+                this.currentDB = newName;
+                this.Text = destiny;
+            }
+
         }
 
         private void toolStripMenuItem6_Click(object sender, EventArgs e)
@@ -144,24 +194,29 @@ namespace SMDB
             Application.Exit();
         }
 
-        private void MenuUpdate(bool enable)
+        public void MenuUpdate(bool enable)
         {
+            /**Desactiva o activa los botones dependiendo de si hay una 
+             * actual base de datos o no
+             * */
             if (enable)
             {
                 //Menu update
                 MenuDisconnectDB.Enabled = true;
-                MenuClose.Enabled = true;
+                MenuHDisconnectDB.Enabled = true;
+                MenuHDeleteDB.Enabled = true;
                 MenuDeleteDB.Enabled = true;
             }
             else
             {
                 //Menu update
                 MenuDisconnectDB.Enabled = false;
-                MenuClose.Enabled = false;
+                MenuHDisconnectDB.Enabled = false;
                 MenuDeleteDB.Enabled = false;
+                MenuHDeleteDB.Enabled = false;
             }
         }
-        private void TitleUpdate()
+        public void TitleUpdate()
         {
             this.Text = this.currentPath;
         }
@@ -181,25 +236,83 @@ namespace SMDB
             CreateDatabase();
         }
 
-        private void DisplayDB()
+        public void DisplayDB()
         {
 
             //Display DB
-            MenuCurrentDB.Items.Clear();
-            MenuCurrentDB.Items.Add(this.currentDB, SMDB.Properties.Resources.severaldb);
-
             //Display all tables
+            TreeView.Nodes.Clear();
             string[] files = System.IO.Directory.GetFiles(this.currentPath);
-            List<string> diccionaries = new List<string>();
-            foreach (string file in files)
+            string[] tables = Array.FindAll(files, x => x.Split('.')[1] == "dbtable");
+            
+            TreeNode[] childs = new TreeNode[tables.Length];
+            for (int i = 0; i < tables.Length; i++)
             {
-                string[] arrFiles = file.Split('.');
-                if (arrFiles[1] == "dd")
-                {
-                    diccionaries.Add(file);
-                }
+                string table_name = tables[i].Split('\\')[tables[i].Split('\\').Length - 1].Split('.')[0];
+                childs[i] = new TreeNode(table_name,0,0);
             }
+            TreeNode rootDatabase = new TreeNode(this.currentDB,10,10,childs);
+            rootDatabase.ContextMenuStrip = MenuRightTreeView;
+            TreeView.Nodes.Add(rootDatabase);
+            TreeView.ExpandAll();
 
+        }
+
+        private void MenuRCreateTable_Click(object sender, EventArgs e)
+        {
+            //Create table
+            this.Create_Table = new CREATE_TABLE();
+            this.Create_Table.Parent_SMBD = this;
+            this.Create_Table.Show();
+            
+        }
+
+        private void MenuHConnectDB_Click(object sender, EventArgs e)
+        {
+            OpenDatabase();
+        }
+
+        private void MenuHDisconnectDB_Click(object sender, EventArgs e)
+        {
+            CloseDatabase();
+        }
+
+        private void MenuHNewDB_Click(object sender, EventArgs e)
+        {
+            CreateDatabase();
+        }
+
+        private void MenuClose_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void MenuHDeleteDB_Click(object sender, EventArgs e)
+        {
+            DeleteDatabase();
+        }
+
+        private void MenuDeleteDB_Click(object sender, EventArgs e)
+        {
+            DeleteDatabase();
+        }
+
+        private void MenuRRenameDB_Click(object sender, EventArgs e)
+        {
+            
+        }
+
+        private void TreeView_BeforeLabelEdit(object sender, NodeLabelEditEventArgs e)
+        {
+            if (e.Node.Level != 0)
+            {
+                e.CancelEdit = true;
+            }
+        }
+
+        private void TreeView_AfterLabelEdit(object sender, NodeLabelEditEventArgs e)
+        {
+            RenameDatabase(e.Label);
         }
     }
 }
